requirements.txt
Django==5.2.7
Pillow==12.0.0
djangorestframework==3.15.2

